+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changes are ordered reverse-chronologically.
 ---
  - Update all user logging related timestamps to a custom datetime field that includes timezone info
  - Newly imported channels are given a 'last_updated' timestamp
+ - Add progress annotation for topics, lazily loaded to increase page load performance
 
 
 0.4